feat: implement add-on feature for unauthenticated users
This commit enables unauthenticated users to add add-ons to products in their cart.
For authenticated users, this feature was already implemented in an earlier commit (153e55a).

Additionally, the `Customer` and `OrderItem` instances has been moved inside the authenticated user check.
This is to prevent potential issues as unauthenticated users intentionally does not have customer and orderitem model instances.