fix: handle `KeyError` for missing `list-of-addons` in cart price calculation
This commit addresses the `KeyError` that occurs when an unauthenticated user initially adds a product to cart but doesn't associate any addons with it.

The key `list-of-addons` is only added to the session cart data when a user associates any addons to any product in cart.
So this causes absence of `list-of-addons` when an unauthenticated user access initially that leads to the `KeyError`.

To resolve this issue, i have modifed the cart price calculation logic.
Now, the total addon price will be considered as 0  if the `list-of-addons` is missing.
Otherwise, iterate  through each product in the cart and their respective addons to sum up the prices.